:bug: fixed simulation bug
when calling simulation_fit(), there was a R-side exception if there were no Zero-Inflated genes. This is caused because the bigmemory matrix descriptor file needs to contain at least one entry. It was corrected by skipping the R-side recalculation of compute_criteria for Zero-Inflated genes (binarisation criteria are used for simulation).