:building_construction: Isolate exception handling on R side
When an exception was raised within the parallel foreach loop in compute_criteria, the child R workers were not terminated. This was solved by making the parallel cluster object a global via the "<<-" operator. Then the cluster was stopped fromthe python side. Using the tryCatch({}) function, the cleanup after exception handling is done completely on the R side without the use of globals.