:construction: working on RNG reproductibility
The random number generator in biased_simulation_from_binary_state() is now fully reproducible. That means that given a fixed seed, the results of calling the function will allways be the same. We would like to ensure that across executions all random number generation runs are fully reproducible, given a fixed seed, for all functions.